pyproject.toml
[tool.poetry]
name = "lpu3dnet"
version = "0.1.0"
description = ""
authors = ["zihan ren"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
torch = {version = "^2.2.0"}
torchvision = {version = "^0.17.0"}
torchaudio = {version = "^2.2.0"}
torchinfo = "^1.8.0"
matplotlib = "^3.8.0"
pandas = "^2.1.0"
scikit-learn = "^1.4.0"
numpy = "^1.24.0"
hydra-core = "^1.3.2"
tqdm = "^4.66.0"
tifffile = "2022.8.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
